Stripped out the horrible, ugly, global dictionary g_dic. All settings now in encoder.yaml
Removed all references to g_dic, now grabbing from settings dictionary. Some
commands now split into multiple lines for easier reading.

Added tempfile to imports, as it always should have been there (though the
function it is used in may be better suited as an outside namespace, since
ParallelEncoding.py uses it too)

Removed move_wr_bits, as it is not really a thing this should do. This is a
little bit too specific to my use (as I dump it in dropbox for auto-sync).
This can be trivially handled with standard low complexity scripting by
the end user, unlike the rest.