Add Encode Versions
Add the ability to automatically insert EPNUMvVERSION in case a new encode
is needed. This is probably most useful for SD encodes following a script
update, but not exclusively.

Also this hunk of the code adds the ability to mux mp4 chapters into a SD
encode, as they've been generated for months but not actually used anywhere.
This requires a patched x264 such as JEEB's, so it's best not to generate
them at all if you don't have it.

Assuming I didn't mess up the --patch hunk selection, this should also
add functionality for default template names so you don't need to type
them every time, though honestly I bat scripted that so long ago I'd
never notice.