Flexible Bitdepth Options for Lossless Pass
It occurred to me after working on the encode_process file that
it was silly to start there. My GARO workflow issues, and probably
anyone else's, are that perhaps the source is only 8-bit, but
you want to use the 10-bit encoder for the lossless pass.

I'm sure I'm missing something here, but now source and encode
bit depths are tracked separately, so the input-depth can be
10 or 16 using the dither hack, or just not set (8) if not.

Additionally, the encoder is set purely by encode-depth, not
source-depth.