Merge from Anemone:
Apparently I made some changes on this file down on the encoder box
as well as the ones to encode_process, and forgot to merge them back
before starting the new dev.

Adds a new function for using sapikachu's hacked ffms2 instead of
TheFluff's unmaintained one. The interface for turning the hack
on differs between the two.

Also, this merges the newer developed changes (source/enc depth)
against those, as the stuff changed elsewhere conflicts.