Added ParallelEncoding-10.py
It is a dumb version of ParallelEncoding.py that only ever does 10-bit and doesn't ever consider anything else, because I never really tested the encode depth switches on the smarter version, and this is the in-use branch.

Also copying up whatever random changes were made to keep shit actually running. I don't even know what they are anymore to be honest, but some of it probably should be merged back to dev. I don't think the -10.py really should be though.