Start actually making the config handle depth
I built the source/enc depth stuff in god knows how long ago now, but
it's never actually done anything of value, and was not in fact
reasonable. As such, it's been switched to a series of in/out depths by
encode type. Also added lossless to the dealio, even though that step
doesn't actually read into the series config section. It should be
cleaner now.