Cleanup shell exec
Clearly I don't need the stupid temporary bat file anymore, but since
all instances where the command is being run with shell=True are going
to need backslashes where python usually doesn't care, I moved the
replacement into the caller itself.

It occurred to me a while back, and now yet again that this shit is
making it wildly unfriendly to linux users wanting to encode with wine
(of which I'm sometimes one) but frankly that's kind of a weird niche
case that I'm not totally sure I care about. Also you should probably
just be looking into vapoursynth for that anyway.