Now with less lingering stupidity
Moved die() declaration above the first use, because it the import yaml try/catch was using it too soon.

Closed parenthesis on the line where I was forcing in a linesep like a putz.

Searching for *ource instead of *source to avoid needing to go case-insensitive, which is kind of dumb now that I type it out in a commit msg.